move function to util folder/package and make dollar sign more readable
