+++ util/find_dollar_sign_test.go
@@ -1,4 +1,4 @@
-package linebot
+package util
 
 import (
 	"testing"
